# src/main/resources/openapi/documentation.yaml
openapi: "3.0.3"
info:
  title: "critica_backend API"
  description: "critica_backend API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
  - url: "https://seahorse-app-qs7ri.ondigitalocean.app"
  - url: "https://critika.net"
paths:
...
  /api/lobby/{id}:
    get:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Get lobby by id"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LobbyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/lobby/{id}/addGame:
    put:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Add game"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "time"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LobbyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/lobby/{id}/addPlayer:
    put:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Add player by name"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LobbyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/lobby/{id}/addPlayer/{playerId}:
    put:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Add player to lobby by id"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "playerId"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LobbyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/lobby/{id}/addTemporaryPlayer:
    put:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Add player without user registration. Note: This endpoint is not secure and should only be used for testing purposes."
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LobbyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/lobby/{id}/delete:
    put:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Delete a lobby"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "204":
          description: "No Content"
          content:
            application/json:
              schema:
                type: "object"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/lobby/{id}/games:
    get:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Get all games in a lobby"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/GameResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/lobby/{id}/players:
    get:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Get players that are currently in the lobby"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PlayerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/lobby/{id}/removeGame/{gameId}:
    put:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Removes a game from the lobby. If the game is not in the lobby, nothing happens."
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "gameId"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LobbyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/lobby/{id}/removePlayer:
    put:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Remove player by name. Only the lobby owner can remove players."
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LobbyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/lobby/{id}/removePlayer/{playerId}:
    put:
      tags:
        - "Lobby"
      security:
        - AccessToken: [ ]
      description: "Remove a player from a lobby. Only the lobby owner can remove players."
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "playerId"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LobbyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/lobby/list:
    get:
      tags:
        - "Lobby"
      security:
        - AccessToken: []
      description: "Get list of lobbies"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/LobbyResponse"
...
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    FirebaseToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PlayerResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        seat:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        role:
          type: "string"
          enum:
            - "CITIZEN"
            - "MAFIOSO"
            - "DON"
            - "DETECTIVE"
        status:
          type: "string"
          enum:
            - "INGAME"
            - "ALIVE"
            - "DEAD"
            - "REMOVED"
            - "OPW"
        inGame:
          type: "boolean"
        bonusPoints:
          type: "integer"
          format: "int32"
...
    DayVoteResponse:
      type: "object"
      properties:
        dayNumber:
          type: "integer"
          format: "int32"
        target:
          $ref: "#/components/schemas/PlayerResponse"
        player:
          $ref: "#/components/schemas/PlayerResponse"
    DayStageResponse:
      type: "object"
      properties:
        type:
          type: "string"
          enum:
            - "DAY"
        dayNumber:
          type: "integer"
          format: "int32"
        candidates:
          type: "array"
          items:
            $ref: "#/components/schemas/PlayerResponse"
        votes:
          type: "array"
          items:
            $ref: "#/components/schemas/DayVoteResponse"
    NightStageResponse:
      type: "object"
      properties:
        dayNumber:
          type: "integer"
          format: "int32"
        mafiaShot:
          $ref: "#/components/schemas/PlayerResponse"
        detectiveCheck:
          $ref: "#/components/schemas/PlayerResponse"
        donCheck:
          $ref: "#/components/schemas/PlayerResponse"
    GameResponse:
      type: "object"
      properties:
        id:
          type: "string"
        date:
          type: "string"
        players:
          type: "array"
          items:
            $ref: "#/components/schemas/PlayerResponse"
        currentStage:
          anyOf:
            - title: "DayStageResponse"
              $ref: "#/components/schemas/DayStageResponse"
            - title: "NightStageResponse"
              $ref: "#/components/schemas/NightStageResponse"
        nominates:
          type: "array"
          items:
            $ref: "#/components/schemas/PlayerResponse"
        votes:
          type: "array"
          items:
            $ref: "#/components/schemas/DayVoteResponse"
        mafiaShot:
          $ref: "#/components/schemas/PlayerResponse"
        detectiveCheck:
          $ref: "#/components/schemas/PlayerResponse"
        donCheck:
          $ref: "#/components/schemas/PlayerResponse"
        playersEliminated:
          type: "array"
          items:
            $ref: "#/components/schemas/PlayerResponse"
    LobbyResponse:
      type: "object"
      properties:
        id:
          type: "string"
        date:
          type: "string"
        name:
          type: "string"
        creator:
          type: "string"
...
    Error:
      type: "object"
      properties:
        status:
          type: "integer"
          format: "int32"
          example: 400
        message:
          type: "string"
          description: "A description of the error."
          example: "Invalid ID"
      required:
        - "status"
        - "message"
  responses:
    BadRequest:
      description: "Malformed data or otherwise invalid request"
      schema:
        $ref: "#/components/schemas/Error"
    NotFound:
      description: "The requested component was not found"
      schema:
        $ref: "#/components/schemas/Error"
    InternalServerError:
      description: "An internal server error occurred"
      schema:
        $ref: "#/components/schemas/Error"
    Forbidden:
      description: "The user does not have the required permissions to access this resource"
      schema:
        $ref: "#/components/schemas/Error"
    Unauthorized:
      description: "The user is not authenticated"
      schema:
        $ref: "#/components/schemas/Error"